refactor(Amount): 로또 구매 금액 자료형 long으로 변경

+++ src/main/java/lotto/controller/LottoController.java
@@ -67,7 +67,7 @@ public class LottoController {
   
   private PurchasedLottos getPurchasedLottos (PurchaseAmount purchaseAmount) {
     int price = lottoService.getPrice();
-    int count = amountService.getPurchasableCount(purchaseAmount, price);
+    long count = amountService.getPurchasableCount(purchaseAmount, price);
     return lottoService.publishPurchaseLottos(count);
   }
